requirements.txt
# Core dependencies 
python-dotenv
supabase

# OpenAI-compatible providers (Phase 0 & Phase 2)
openai
groq
together
openrouter

# Custom SDK providers (Phase 3)
anthropic
google-generativeai

# Utilities
httpx
requests
beautifulsoup4

# Local database (PostgreSQL)
psycopg2-binary
